+++ src/Objects/DeviceSound.php
@@ -20,7 +20,7 @@ class DeviceSound
     
     public function __construct(
         public string $time,
-        public float $value,
+        public int $value,
     ) {
     }
