+++ src/Objects/DeviceTemperature.php
@@ -20,7 +20,7 @@ class DeviceTemperature
     
     public function __construct(
         public string $time,
-        public float $value,
+        public int $value,
     ) {
     }
